refactor(convertFileToGeojson): convertFileToGeojson function add to utilFunctions
