+++ src/frontend/src/components/createnewproject/UploadArea.tsx
@@ -15,6 +15,7 @@ import { valid } from 'geojson-validation';
 import useDocumentTitle from '@/utilfunctions/useDocumentTitle';
 import DescriptionSection from '@/components/createnewproject/Description';
 import UploadAreaComponent from '@/components/common/UploadArea';
+import { convertFileToGeojson } from '@/utilfunctions/convertFileToGeojson';
 
 type uploadAreaType = 'upload_file' | 'draw';
 
@@ -86,27 +87,6 @@ const UploadArea = ({ flag, geojsonFile, setGeojsonFile, setCustomDataExtractUpl
     dispatch(CreateProjectActions.SetToggleSplittedGeojsonEdit(false));
   };
 
-  const convertFileToGeojson = async (file) => {
-    if (!file) return;
-    const fileReader = new FileReader();
-    const fileLoaded: any = await new Promise((resolve) => {
-      fileReader.onload = (e) => resolve(e.target?.result);
-      fileReader.readAsText(file, 'UTF-8');
-    });
-    const parsedJSON = JSON.parse(fileLoaded);
-    let geojsonConversion;
-    if (parsedJSON.type === 'FeatureCollection') {
-      geojsonConversion = parsedJSON;
-    } else {
-      geojsonConversion = {
-        type: 'FeatureCollection',
-        features: [{ type: 'Feature', properties: null, geometry: parsedJSON }],
-      };
-    }
-    addGeojsonToState(geojsonConversion);
-    return geojsonConversion;
-  };
-
   const addGeojsonToState = (geojson) => {
     dispatch(CreateProjectActions.SetDrawnGeojson(geojson));
   };
@@ -118,10 +98,11 @@ const UploadArea = ({ flag, geojsonFile, setGeojsonFile, setCustomDataExtractUpl
     }
 
     const fileObject = file?.file;
-    if (valid(await convertFileToGeojson(fileObject))) {
+    const convertedGeojson = await convertFileToGeojson(fileObject);
+    if (valid(convertedGeojson)) {
       handleCustomChange('uploadedAreaFile', fileObject?.name);
       setGeojsonFile(file);
-      convertFileToGeojson(fileObject?.file);
+      addGeojsonToState(convertedGeojson);
     } else {
       handleCustomChange('uploadedAreaFile', '');
       setGeojsonFile(null);
